requirements.txt
python-telegram-bot==20.3
pydub==0.25.1
ffmpeg-python==0.2.0
python-dotenv==1.0.0